Update README.md
The goal can be easily achieved by using the length unction which makes the program easy. The number in the array can be checked by taking the percentile. If the remainder is zero then it i a even number otherwise it is a odd number. The print function is used to print the number which passes the if condition. The numbers which go into the if loop are getting printed.